pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "openenv-reddit-mod-env"
version = "1.0.0"
description = "Reddit Moderator Bot RL environment for the OpenEnv hackathon — train agents to moderate r/science"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["reddit_mod_env*"]

[tool.setuptools.package-dir]
"" = "."
